feat: structured-packages-implementation-auto-02-cron-001 - Own cron behind the Automations root
