+++ docs/internal/baselines/go-functional-coverage-package-minimums.json
@@ -115,6 +115,24 @@
       }
     },
     {
+      "package": "github.com/portpowered/infinite-you/pkg/services/automations/internal/services/cron",
+      "exception": {
+        "kind": "measurement",
+        "justification": "The active functional coverage profile contains no measurable statements for this package.",
+        "owner": "backend-quality",
+        "deadline": "2027-07-15",
+        "removalGate": "The functional coverage profile reports at least one measurable statement for this package."
+      }
+    },
+    {
+      "package": "github.com/portpowered/infinite-you/pkg/services/automations/internal/services/cron/internal/service",
+      "minimum": 63.63
+    },
+    {
+      "package": "github.com/portpowered/infinite-you/pkg/services/automations/internal/services/cron/wire",
+      "minimum": 0.00
+    },
+    {
       "package": "github.com/portpowered/infinite-you/pkg/services/automations/internal/services/reconciliation",
       "exception": {
         "kind": "measurement",
@@ -135,10 +153,6 @@
     {
       "package": "github.com/portpowered/infinite-you/pkg/services/automations/service",
       "minimum": 25.94
-    },
-    {
-      "package": "github.com/portpowered/infinite-you/pkg/services/automations/timework",
-      "minimum": 63.63
     },
     {
       "package": "github.com/portpowered/infinite-you/pkg/services/edges",
