+++ docs/internal/baselines/go-unit-coverage-package-minimums.json
@@ -115,6 +115,24 @@
       }
     },
     {
+      "package": "github.com/portpowered/infinite-you/pkg/services/automations/internal/services/cron",
+      "exception": {
+        "kind": "measurement",
+        "justification": "The active unit coverage profile contains no measurable statements for this package.",
+        "owner": "backend-quality",
+        "deadline": "2027-07-15",
+        "removalGate": "The unit coverage profile reports at least one measurable statement for this package."
+      }
+    },
+    {
+      "package": "github.com/portpowered/infinite-you/pkg/services/automations/internal/services/cron/internal/service",
+      "minimum": 83.83
+    },
+    {
+      "package": "github.com/portpowered/infinite-you/pkg/services/automations/internal/services/cron/wire",
+      "minimum": 100.00
+    },
+    {
       "package": "github.com/portpowered/infinite-you/pkg/services/automations/internal/services/reconciliation",
       "exception": {
         "kind": "measurement",
@@ -135,10 +153,6 @@
     {
       "package": "github.com/portpowered/infinite-you/pkg/services/automations/service",
       "minimum": 81.08
-    },
-    {
-      "package": "github.com/portpowered/infinite-you/pkg/services/automations/timework",
-      "minimum": 83.83
     },
     {
       "package": "github.com/portpowered/infinite-you/pkg/services/edges",
