+++ docs/internal/baselines/package-structure-baseline.json
@@ -2509,11 +2509,6 @@
     },
     {
       "rule": "service-root-unexpected-directory",
-      "filePath": "pkg/services/automations/timework",
-      "target": "pkg/services/automations"
-    },
-    {
-      "rule": "service-root-unexpected-directory",
       "filePath": "pkg/services/factory_definitions/authoredlayout",
       "target": "pkg/services/factory_definitions"
     },
