+++ pkg/services/automations/service/service.go
@@ -9,6 +9,8 @@ import (
 	"github.com/jonboulle/clockwork"
 	automations "github.com/portpowered/infinite-you/pkg/services/automations"
 	reconciliation "github.com/portpowered/infinite-you/pkg/services/automations/internal/services/reconciliation"
+	cron "github.com/portpowered/infinite-you/pkg/services/automations/internal/services/cron"
+	cronwire "github.com/portpowered/infinite-you/pkg/services/automations/internal/services/cron/wire"
 	factorydefinitions "github.com/portpowered/infinite-you/pkg/services/factory_definitions"
 	"github.com/portpowered/infinite-you/pkg/services/workers"
 	"go.uber.org/zap"
@@ -32,6 +34,7 @@ type Service struct {
 	resolveTemplates  workers.TemplateFieldResolver
 	executionPolicy   factorydefinitions.WorkstationExecutionPolicyService
 	reconciler        reconciliation.Service
+	cron              cron.Service
 	schedulerMu       sync.Mutex
 	schedulerSources  map[automations.SourceIdentity]*schedulerSource
 }
@@ -60,6 +63,7 @@ func New(
 		schedulerSources:  make(map[automations.SourceIdentity]*schedulerSource),
 	}
 	service.reconciler = service.newSchedulerReconciler()
+	service.cron = cronwire.NewService()
 	return service
 }
